% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make}
\alias{make}
\title{Shiny app launcher}
\usage{
make()
}
\value{
A .R file which correspond to the converted function.
}
\description{
Launch the shiny app that convert the user function to a safer-r function.
}
\author{
\href{mailto:gael.millot@pasteur.fr}{Gael Millot}
}
